requirements.txt
# --- Core API + Server ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- LangChain + OpenAI + Pinecone ---
langchain==0.2.14
langchain-openai==0.1.22
tiktoken==0.7.0
openai==1.43.0
pinecone-client==5.0.1

# --- Config & Environment ---
pydantic-settings==2.4.0
python-dotenv==1.0.1

# --- PDF & HTML Parsing ---
pypdf==4.3.1
beautifulsoup4==4.12.3
markdown==3.6

# --- Logging & Debugging ---
rich==13.8.1
loguru==0.7.2        # Advanced logging

# --- Google Drive Integration ---
google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# --- Scheduling / Background Tasks ---
apscheduler==3.10.4  # For periodic checks
celery==5.4.0        # If you want distributed tasks (optional)

# --- Async Helpers ---
aiofiles==24.1.0      # For async file handling
httpx==0.27.2         # Async HTTP requests if needed
apscheduler